+++ collector-amazon-ivs/src/test/java/com/bitmovin/analytics/amazon/ivs/IAmazonIvsPlayerCollectorKotlinTest.kt
@@ -1,0 +1,19 @@
+package com.bitmovin.analytics.amazon.ivs
+
+import com.bitmovin.analytics.BitmovinAnalyticsConfig
+import io.mockk.mockk
+import org.junit.Test
+
+class IAmazonIvsPlayerCollectorKotlinTest {
+
+    // This test is a sanity check that the kotlin factory stays stable and consistent with the naming
+    @Test
+    fun testFactory_shouldCreateNewCollectorObject() {
+        // arrange
+        val config = BitmovinAnalyticsConfig("test-analytics-key", "test-player-key")
+
+        // act and assert (since return value is non nullable we get a valid object or an exception with this call)
+        val collector = IAmazonIvsPlayerCollector.Factory.create(config, mockk(relaxed = true))
+        val collector2 = IAmazonIvsPlayerCollector.create(config, mockk(relaxed = true))
+    }
+}