+++ collector-amazon-ivs/src/test/java/com/bitmovin/analytics/amazon/ivs/TestUtils.kt
@@ -1,9 +1,11 @@
 package com.bitmovin.analytics.amazon.ivs
 
+import android.content.Context
 import com.bitmovin.analytics.data.DeviceInformation
 import com.bitmovin.analytics.data.EventData
 import com.bitmovin.analytics.data.PlayerInfo
 import com.bitmovin.analytics.enums.PlayerType
+import io.mockk.mockk
 
 class TestUtils {
     companion object {
@@ -34,6 +36,11 @@ class TestUtils {
                 null, null, null, null, null,
                 null, null, "testUserAgent",
             )
+        }
+
+        // helper to create a context mock for the java tests
+        fun createMockContext(): Context {
+            return mockk(relaxed = true)
         }
     }
 }