+++ collector-bitmovin-player/src/test/java/com/bitmovin/analytics/bitmovin/player/TestUtils.kt
@@ -1,0 +1,13 @@
+package com.bitmovin.analytics.bitmovin.player
+
+import android.content.Context
+import io.mockk.mockk
+
+class TestUtils {
+    companion object {
+        // helper to create a context mock for the java tests
+        fun createMockContext(): Context {
+            return mockk(relaxed = true)
+        }
+    }
+}